Basics of breaking up LiveAuctionViewController.swift.
