+++ Artsy/View_Controllers/Live_Auctions/LiveAuctionSaleViewController.swift
@@ -1,0 +1,17 @@
+import UIKit
+
+/// A class to display a fetched and fully-populated LiveSaleViewModel. Hosted by LiveAuctionViewController, which does the fetching.
+class LiveAuctionSaleViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+}
